+++ src/main/java/com/pickkasso/domain/curriculum/dto/response/AllCurriculumListViewResponse.java
@@ -1,4 +1,4 @@
-package com.pickkasso.domain.curriculum.dto;
+package com.pickkasso.domain.curriculum.dto.response;
 
 import com.pickkasso.domain.usercurriculum.domain.StateType;
